backend/requirements.txt
Flask
Pillow
werkzeug
opencv-python
numpy
redis
python-dotenv